+++ src/main/kotlin/com/lykke/matching/engine/fee/FeeCoefCalculator.kt
@@ -1,5 +1,7 @@
 package com.lykke.matching.engine.fee
 
+import java.math.BigDecimal
+
 interface FeeCoefCalculator {
-    fun calculate(): Double?
+    fun calculate(): BigDecimal?
 }
